+++ crates/entrypoint/src/cli/command/azure_devops/license_entitlement/user/mod.rs
@@ -1,8 +1,10 @@
 pub mod list;
 pub mod update;
+pub mod show;
 
 pub use list::AzureDevOpsLicenseEntitlementUserListArgs;
 pub use update::{
     AzureDevOpsLicenseEntitlementUserUpdateArgs,
     AzureDevOpsLicenseEntitlementUserUpdateTuiArgs,
 };
+pub use show::AzureDevOpsLicenseEntitlementUserShowArgs;