+++ crates/entrypoint/src/cli/command/azure_devops/license_entitlement/user/show/mod.rs
@@ -1,0 +1,2 @@
+mod azure_devops_license_entitlement_user_show_cli;
+pub use azure_devops_license_entitlement_user_show_cli::*;